[refactor] move tests from /tests to /actions/order
